chore: change GoalTopBar Component
